Refactor of ViafClient and tests to turn ViafClient methods to static. This involved altering tests so that it used PowerMock (specifically PowerMockito) so that we can mock and test static methods.
As a result of needing to use PowerMockito mockito was downgraded to v1.9.0 as the latest released version of PowerMock was not compatible with v.1.9.5. I recommend keeping an eye on PowerMock and as soon as v1.4.13 is available (should be compatible with Mockito v1.9.5) we upgrade to those versions.

Also downgraded to junit 4.8.2 as 4.10 is not compatible with PowerMockito 1.4.12.